Improved auth flow, added logic for syncing FCM token data with cloid, fixed  logError bug
I improved the auth flow of the app to make it cleaner and faster. It now makes use of AsyncStorage to check if the user has properly set uo the account, allowing the user to actually access the full app  while offline upon launch.

I alos fixed an error causing typo with my logError fuction.

I also added logic for updating the cloud with data about the user's FCM token. This also makes use of AsyncStorage