Added yarn scripts for building for prod or dev environments
I also had to change the way I handle the two different environments in ios. A strange error in Xcode made it unable to recognize a new plist file that was created in the build phases during building. So, made sh files to copy the right plist file before building even starts.